Begin adding tests! Start with the Dice plugin.
This test spec was instrumental in fixing the formatting issues
introduced a few commits ago. Thanks go to samphippen in the
Freenode #rspec channel for answering the newbie questions that
made this possible. The sheer volume of problems that could have
been caught by a decent test suite is insane. I get it now.